refactor(tests): update string empty checks and improve test case consistency
Refactored multiple test cases to replace System.String.Empty with "" for better readability and consistency. Adjusted related test cases to ensure they accurately reflect the intended behavior.

No functional changes were made to the logic.

+++ ConstExpr.Tests/Tests/Rewriter/VisitCastExpressionTests.cs
@@ -18,7 +18,7 @@ public class VisitCastExpressionTests : BaseTest<Func<double, int, int, (int, do
 
 	public override IEnumerable<KeyValuePair<string?, object?[]>> TestCases =>
 	[
-		Create((x, y, z) => ((int)x, y, (char)z, 3)),
+		Create((x, y, z) => ((int)x, (double)y, (char)z, 3)),
 		Create((_, _, _) => (3, 42.0, 'A', 3), [ 3.14, 42, 65 ]),
 		Create((_, _, _) => (10, 100.0, 'Z', 3), [ 10.5, 100, 90 ]),
 		Create((_, _, _) => (-5, -10.0, ' ', 3), [ -5.8, -10, 32 ])
